# sms-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "School Management System — Go + Gin + PostgreSQL\n\n## Authentication\nThis API uses JWT Bearer tokens. Login via `POST /api/login` to receive an access token (1 hour) and a refresh token (7 days).\nClick the **Authorize** button above and enter: `Bearer \u003cyour_token\u003e`\n\n## Roles\n| Role    | Access                                              |\n|---------|-----------------------------------------------------|\n| Admin   | Full access to all endpoints                        |\n| Teacher | Attendance, grades, locker (read public)             |\n| Student | Own grades, report card, locker, finance             |\n| Parent  | Children's data, report cards, finance               |",
        "title": "SMS Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/admin/finance/payroll": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a payroll record for a teacher for a specific month and year. Admin only. Prevents duplicates.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "finance"
                ],
                "summary": "Create a payroll entry",
                "parameters": [
                    {
                        "description": "Payroll data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CreatePayrollRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Payroll created",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error or teacher not found",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden — Admin only",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    },
                    "409": {
                        "description": "Payroll already exists for this month",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/admin/finance/payroll/{id}/pay": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sets a payroll record's status to Paid and records the payment timestamp. Admin only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "finance"
                ],
                "summary": "Mark payroll as paid",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Payroll ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payroll updated",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Already paid",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden — Admin only",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Payroll not found",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/admin/finance/receipt/{id}/verify": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin marks a pending receipt as Verified or Rejected. Can only be done once per receipt.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "finance"
                ],
                "summary": "Verify or reject a bank receipt",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Verification data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.VerifyReceiptRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Receipt processed",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Already processed or validation error",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden — Admin only",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Transaction not found",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/api/finance/receipt": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Student or Parent submits proof of payment. Duplicate receipt IDs are rejected.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "finance"
                ],
                "summary": "Submit a bank receipt",
                "parameters": [
                    {
                        "description": "Receipt data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.SubmitReceiptRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Receipt submitted",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden — Student/Parent only",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    },
                    "409": {
                        "description": "Receipt ID already exists",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/finance/transactions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin sees all transactions. Students see only their own. Parents see transactions for all their linked children.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "finance"
                ],
                "summary": "List transactions",
                "responses": {
                    "200": {
                        "description": "Transaction list",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Student/user profile not found",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.CreatePayrollRequest": {
            "type": "object",
            "required": [
                "amount",
                "month",
                "teacher_id",
                "year"
            ],
            "properties": {
                "amount": {
                    "type": "number",
                    "minimum": 1,
                    "example": 15000
                },
                "month": {
                    "type": "integer",
                    "maximum": 12,
                    "minimum": 1,
                    "example": 5
                },
                "teacher_id": {
                    "type": "integer",
                    "example": 2
                },
                "year": {
                    "type": "integer",
                    "example": 2025
                }
            }
        },
...
        "controllers.SubmitReceiptRequest": {
            "type": "object",
            "required": [
                "amount",
                "receipt_id",
                "student_id"
            ],
            "properties": {
                "amount": {
                    "type": "number",
                    "minimum": 1,
                    "example": 5000
                },
                "description": {
                    "type": "string",
                    "example": "Term 1 tuition fee"
                },
                "receipt_id": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 5,
                    "example": "CBE-TXN-20250501"
                },
                "student_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "controllers.VerifyReceiptRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "remarks": {
                    "type": "string",
                    "example": "Confirmed with bank"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "Verified",
                        "Rejected"
                    ],
                    "example": "Verified"
                }
            }
        },
        "helpers.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and your JWT token. Example: **Bearer eyJhbGci...**",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}